Added Hvvp class which is identical to Paul's Hvvp code, and compiles, but is currently not implemented in the algorithm because it requires additional computed parameters, so it has not been tested either.  Added Lamb and Deformation Field analytic model options in AnalyticGrid and vortrac_defaultAnalyticTC.xml, these are still being tested.

+++ trunk/Radar/Ray.cpp
@@ -148,11 +148,11 @@ int Ray::getFirst_vel_gate() {
   return first_vel_gate ;
 }
 
-int Ray::getRef_gatesp() {
+float Ray::getRef_gatesp() {
   return ref_gatesp ;
 }
 
-int Ray::getVel_gatesp() {
+float Ray::getVel_gatesp() {
   return vel_gatesp ;
 }
